fix api mem leak due to validator profiles
